# [assets] Smaller bundles and remove React dev code.