--- a/scripts/assets.sh
+++ b/scripts/assets.sh
@@ -8,6 +8,9 @@ mkdir public/assets
 NODE_ENV=production browserify \
   $(find desktop/assets mobile/assets -name '*.coffee') \
   -p prundupify \
+  -g [ envify --NODE_ENV production ] \
+  -g uglifyify \
+  -p bundle-collapser/plugin \
   -t babelify \
   -t caching-coffeeify \
   -t jadeify \