@alloy 2017-05-10
Offtopic: but I can’t find any source maps for these bundles, where/how is that being handled?