@craigspaeth 2017-05-10
Nice, thanks for endeavoring on this! I see React's docs _do_ suggest envify—sorry about that @damassi I swore it only required `NODE_ENV=production` in the past. Source maps can be enabled in browserify via the `--debug` flag and [extracted via a plugin](https://github.com/substack/node-browserify#external-source-maps).