@alloy 2017-05-10
Cool, thanks, I’ll look at source maps from browserify as part of my compiled Force branch 👍