@damassi 2017-05-10
Ah! apologies -- `livereload` is like a pre-HMR technology where after something compiles it will send a reload request to the browser so that devs don't have to manually refresh every time something changes. I've been thinking about adding something like this https://www.npmjs.com/package/browserify-livereload but have been bogged down with other work.