@alloy 2017-05-10
Oh I see. I’d feel more comfortable making such changes if I were actually using it, to be honest.

Also, if it’s _just_ reloading a window, it kinda sounds like you could just use any ol’ file watcher, no? E.g. using kicker: `kicker -e 'open http://localhost:5000/foo/bar' desktop/apps/auctions`