@damassi 2017-05-10
Totally -- was more of a wink than a request 😄  It's something I've been wanting to add for a little bit to make things a bit quicker in Force.